fix-shellcheck(SC1128) The shebang must be on the first line. (#815)
The shebang must be on the first line. Delete blanks and move comments.shellcheck(SC1128) #814

A shebang only has an effect when it appears as the first line in a script. Specifically, the first two bytes of the file must be #!.

More external link see <https://www.shellcheck.net/wiki/SC1128>